requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=py38h06a4308_0
aedat=1.2.2=pypi_0
aiohttp=3.7.4=py38h27cfd23_1
apex=0.1=pypi_0
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=py38h497a2fe_2
async-timeout=3.0.1=py38h06a4308_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
blessings=1.7=pypi_0
blinker=1.4=py38h06a4308_0
blosc=1.21.0=h9c3ff4c_0
blosc-hdf5-plugin=1.0.0=h646ab9b_2
bottleneck=1.3.2=py38heb32a55_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=3.0.4=py38h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py38_0
cython=0.29.24=py38h295c915_0
cytoolz=0.11.0=py38h497a2fe_3
dask-core=2021.7.2=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
dense-deep-event-stereo=1.0=dev_0
dv=1.0.10=pypi_0
einops=0.3.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
enum34=1.1.10=py38h32f6830_2
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fitlog=0.9.13=pypi_0
flask=2.0.2=pypi_0
flatbuffers=2.0=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd8ed1ab_0
gitdb=4.0.8=pypi_0
gitpython=3.1.24=pypi_0
glib=2.69.0=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
gpustat=0.4.1=pypi_0
grpcio=1.36.1=py38h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.2.1=py38h6c542dc_0
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imagecodecs-lite=2019.12.3=py38h5c078b8_3
imageio=2.9.0=py_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.5.5=py38hd0cf306_0
ipython=7.22.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.4=pyhd8ed1ab_0
itsdangerous=2.0.1=pypi_0
jedi=0.17.0=py38_0
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py38h578d9bd_6
jupyter_client=7.0.2=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h578d9bd_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.1=pyhd8ed1ab_0
kiwisolver=1.3.1=py38h2531618_0
kornia=0.6.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.37.0=py38he1b5a44_0
locket=0.2.0=py_2
lz4=4.0.0=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
mistune=0.8.4=py38h497a2fe_1004
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
multidict=5.1.0=py38h27cfd23_2
nbclient=0.5.4=pyhd8ed1ab_0
nbconvert=6.1.0=py38h578d9bd_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
netron=5.4.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.3=py_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.3=pyha770c72_0
numba=0.54.0=np1.11py3.8hc13618b_g5888895d3_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py_0
onnx=1.10.2=pypi_0
onnxruntime-gpu=1.9.0=pypi_0
opencv-python=4.5.3.56=pypi_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.2=py38h8c16a72_0
pandoc=2.14.2=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd8ed1ab_0
path=16.0.0=py38h06a4308_0
path.py=12.5.0=0
pathlib=1.0.1=py38h578d9bd_4
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
pooch=1.4.0=pyhd8ed1ab_0
practicaldeepstereo=0.1=dev_0
prettytable=2.2.0=pyhd8ed1ab_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.17.2=py38h295c915_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h497a2fe_2
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.19.1=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.1=py3.8_cuda10.1_cudnn7.6.3_0
pytorch-msssim=0.2.1=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38h5c078b8_3
pyyaml=5.4.1=py38h497a2fe_0
pyzmq=19.0.2=py38ha71036d_2
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd8ed1ab_0
qtpy=1.11.0=pyhd8ed1ab_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.2=py38h1abd341_0
scikit-video=1.1.11=pyh24bf2e0_0
scipy=1.6.2=py38had2a1c9_1
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pyhd3eb1b0_0
terminado=0.12.1=py38h578d9bd_0
testpath=0.5.0=pyhd8ed1ab_0
thop=0.0.31-2005241907=pypi_0
tifffile=2019.7.26.2=py38_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchstat=0.0.7=pypi_0
torchvision=0.9.1=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.61.2=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py_1
werkzeug=2.0.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38h578d9bd_4
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.5.1.0=pypi_0
